+++ src/utils.ts
@@ -1,22 +1,11 @@
 import "url-search-params-polyfill";
-import pluginConfig, { VueGtmUseOptions } from "./config";
+import { VueGtmUseOptions } from "./config";
 
 /**
- * Console log depending on config debug mode
+ * Load GTM script tag.
  *
- * @param message
- */
-export function logDebug(message: string, args: Record<string, any>): void {
-  if (pluginConfig.debug) {
-    console.log("VueGtm :", ...arguments);
-  }
-}
-
-/**
- * Load GTM script tag
- *
- * @param id GTM ID
- * @param params query params object
+ * @param id GTM ID.
+ * @param config The config object.
  */
 export function loadScript(
   id: string,
@@ -49,7 +38,9 @@ export function loadScript(
 }
 
 /**
- * Check if GTM script is in the document
+ * Check if GTM script is in the document.
+ *
+ * @returns `true` if in the `document` is a `script` with `src` containing `googletagmanager.com/gtm.js`, otherwise `false`.
  */
 export function hasScript(): boolean {
   return Array.from(document.getElementsByTagName("script")).some((script) =>